% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{lr}
\alias{lr}
\title{Fitting Linear Regressions}
\usage{
lr(X, Y, I)
}
\arguments{
\item{X}{a matrix or a dataframe contains all covariates of interest as columns
and all observations. If you want your model contains an intercept, the first column
of X should be fulled with 1.}

\item{Y}{the dependent variable. It is a matrix with dimension n*1 (n means
number of observations).}

\item{I}{a logical operator for the intercept. If I = TRUE, the model contains
the intercept, or the intercept is removed.}
}
\value{
\item{coefficients}{a vector for each covariate including estimated
coefficients, standard error, t value, p value and significance}

\item{Residual standard error}{the standard error of residual}

\item{R-squared value}{the percentage of varience can be explained by
the model. Higher percentage means a better fitting of model}

\item{F statistic}{the test statistic for the analysis of variance
approach to test the significance of the model or the components in the model}
}
\description{
This function is used to fit linear models, including both simple linear
regressions and multiple linear regressions. The output includes the estimated
coefficient, standard error, t values, and P values for each variable, and
R squared value, adjusted R squared value, Residual standard error, F-statistic
and p value for the whole model.
}
\examples{
X = iris[,c(2,3)]
Y = iris$Sepal.Length
lr(X,Y,TRUE)
}
